Fix a bug in wit-component when a function is only used within an adapter (#894)
This commit fixes a bug in `wit-component` when used with adapters when
an imported interface or function is used exclusively within an adapter
that otherwise isn't needed by the main module. The fix here was to
alter the order of iteration when calculating shims to iterate over all
adapters instead of only iterating over the adapters required by the
main module.