wit: Add support for type definitions in `world` items
This commit adds support for having a `type` declaration in a `world`
block along with a `use` declaration to import types. This finishes up
the `world` implementation for the upstream specification and makes
interfaces and worlds more uniform by allowing all the types in all the
places.

The implementation here has nontrivial work to do within the AST
resolution phase to make sure everything is sorted out.
Encoding/decoding has also been updated to be generalized in a few more
places in addition to supporting root-level type exports which map to
type definitions. Finally the fuzzer has been updated to generate worlds
that have type definitions in them.