Fix failing `*.wast` tests and align APIs
This commit fixes some decoding of types in the `wasmparser` crate and
then additionally adds support in `wasmparser`, `wasm-encoder`, etc, for
the initialization expression of a table being specified. This involved
aligning the type hierarchy of `wasm-encoder` with that of `wasmparser`
which involved quite a few changes in a number of crates. Overall though
this is mostly syntactic changes without much meat happening here.