Implement export type ascription
This commit implements the change described in
WebAssembly/component-model#155 and accounts for the update in
WebAssembly/component-model#161. The internal plumbing for this change
was relatively easy to implement and shouldn't cause too much of a
burden even on Wasmtime as a consumer (it may not have to change at
all). I'll be working on a follow-up for `wit-component` to use this new
construct to fix the original underlying issue.