Remove function-based subtyping
This commit removes the function-based subtyping support from the
component model validation implementation. This means that subtyping,
when it comes to value types or function types, is effectively a full
type equality check.

The main rationale for this commit is that subtyping is not implemented
in Wasmtime today, it has no current path to becoming implemented, and
there is not current planned use case to require subtyping in the
immediate future. One of the main motivations for subtyping, evolution
of components over time, will have a story that goes beyond subtyping
and will need to account for things like bindings generation as well.

The purpose here is to put Wasmtime in a more conservative position
where subtyping can be added later if necessary but otherwise subtyping
isn't a "hole" in the implementation for now. As-is you can certainly
panic Wasmtime with subtyping and my guess is that you can trigger
memory unsafety as well, but I haven't fully proven that out.
Nonetheless Wasmtime is an incorrect implementation of the component
model with subtyping as-is right now.

It's worth noting, however, that limited subtyping still remains.
Notably instance/component/module subtyping all accounts for "you can
export more things" or "you can import fewer things". This variant of
subtyping is correctly implemented in Wasmtime and is useful in the
real-world today to avoid needing to keep the entire world in perfect
alignment. I wish I had a better rationale for keeping this other than
"well it works" as opposed to function subtyping which is "well it
doesn't work", but that's mostly what I've got right now.