Move the wasmparser-dump crate into the CLI
This commit removes the extraneous `wasmparser-dump` crate which was
only a historical oddity really. Instead now it's located directly in
the CLI itself like other commands are (e.g. objdump). To preserve
testing this commit additionally adds a new "cli" test suite which is
intended to allow testing the CLI itself and its functionality.