Model types-in-worlds as imports
This switches types-in-worlds from being modeled as exports to instead
being modeled as imports. This allows them, in the binary encoding, to
be used for imported functions in addition to exported functions.
Internally this was mostly "just" a change about consulting different
sets and moving things around. This additionally simplified a few
"carefully iterate over items in the right order" loops as imports are
now always topologically before all exports.