Reduce size of `get_rules` function
This commit, like the previous commit, tries to fix a wasm-mutate
function definition which, when compiled to wasm in debug mode,
previously required more than 50k locals. This was done by using a
custom `rewrite!` macro for this method which does the same thing as the
one in the `egg` crate but delegates to local methods to expand to less
inline code related to parsing and such.