Reduce the code size of some wasm-mutate functions (#917)
* Reduce compiled size of WasmMutate::run

I was playing around with compiling the `wasm-tools` CLI to wasm itself
but it unfortunately didn't work in Wasmtime or within `wasm-tools`
itself because some `wasm-mutate` functions have too many locals. Namely
the `WasmMutate::run` function, in debug mode, has over 100k locals.
This commit simplifies the function to not use a macro that generates
O(n^2) code and instead does a bit of dynamic dispatch at runtime.

* Reduce size of `get_rules` function

This commit, like the previous commit, tries to fix a wasm-mutate
function definition which, when compiled to wasm in debug mode,
previously required more than 50k locals. This was done by using a
custom `rewrite!` macro for this method which does the same thing as the
one in the `egg` crate but delegates to local methods to expand to less
inline code related to parsing and such.