Update filter for tag builds (#929)
Looks like if a tag filter is provided to GitHub Actions then when tags
are pushed as a group if any tag doesn't match the filter then CI
doesn't run at all. This means that the mass-push of tags for crates in
this repo accidentally don't run CI to build the release binaries for
the tag of `wasm-tools` itself.

This commit fixes the issue by updating the conditions for the
`build.yml` workflow. This always runs on all tags now but the only job
will be skipped unless it's a `wasm-tools` tag.