wasmparser: fix component validation for type instantiation arguments.
This commit fixes validation when a component is instantiated that imports a
type and the type is provided as an instantiation argument.

Previously, the type was being ignored and not populated in the list of
instantiation arguments, resulting in an incorrect validation message of
"instantiation argument not provided".

The fix is to add type arguments to the arguments list.