wasm-compose: fix incorrect type indexes caused by type imports.
Currently `wasm-compose` writes all core and component types out in
contiguous sections.

The problem with this approach is that imports of types occur *after* these
sections, so an import of a type will be assigned an incorrect type index.

The fix is to simply encode core types, component types, and imports as
encountered rather than trying to group them all into the same section.