Increase the maximum type size to 1M (#945)
In #940 there are modules in the wild running into the previous 100k
limit. This previous limit was completely arbitrarily chosen, so
increase it by 10x and see how things go. In theory 1M should still be
reasonable enough to typecheck in a modest amount of time for input
modules while still preventing excessive use of resources during
fuzzing.