Update support for the relaxed-simd proposal (#944)
This wasm proposal is almost phase 4 so now seemed like a reasonable
time to sync the implementation here in `wasm-tools` with the upstream
specification. The changes here are:

* The "sat" in the float-to-int intrinsics naming was removed since
  it's not present in the upstream instruction.
* Dot-product instructions now have "relaxed" in their name.
  Additionally the float-based dot-product instruction has been removed
  from the upstream proposal.
* Opcodes were audited to ensure they match the latest upstream
  proposal.
* The "fma" and "fnma" naming was updated to "madd" and "nmadd"
* Generators in `wasm-smith` were updated to match the name of the
  instruction.
* The validator was reorganized to put all the relaxed simd instructions
  next to each other.

The upstream spec test suite was also updated which brought in some
BLESS=1 snapshot updates as well along with a minor tweak to an error
message.