feat: preliminary support for WIT templates
This is a minimum viable implementation of WIT templates per
https://github.com/WebAssembly/component-model/issues/172.  It
supports interfaces with at most one wildcard function, which may be
expanded (i.e. monomorphized) with a set of substitutions provided by
the application developer when generating guest bindings.  I will be
posting separate PRs to the `wit-bindgen` and `wasmtime` repos to
implement and test guest and host binding generation, respectively.

Signed-off-by: Joel Dice <joel.dice@fermyon.com>