Elaboration of table initialisation with typeful references
This commit fixes a bug in the elaboration of table initialisation
when the element has a typeful reference type. Previsouly, the
elaborator would assume that the initialising element has type
`funcref`. Concretely, in code this meant that the following module

```wast
(module
  (type $t (func))
  (func $tf)
  (table $t (ref null $t) (elem $tf)))
```

would elaborate to something like

```wast
(module
  (type $t (func))
  (func $tf (type $t))
  (table $t 1 1 (ref null 0))
  (elem 0 (i32.const 0) func (ref.func $tf)))
```

The problem here is that the type of `ref.func $tf` is `ref null 0`
which is not a subtype of `func`. So, now the wast parser correctly
elaborates the above module, such that the it reads `(ref null 0)`
rather than `func` in the type component of `elem`.

The current implementation uses a slight hack to back patch a parsed
`ElemPayload` as the parser do not have access to the sufficient
context information to parse `ElemPayload` correctly for the above
example.