Fix base case for wit-parser's variant alignment
This fixes a bug in #948 discovered by wit-bindgen's test suite which I
forgot to run prior to landing it here. Historically much of the ABI
code in `wit-parser` was tested by `wit-bindgen` but after the
extraction there is no testing of it in this repository, so this is
something we'll want to improve in the future.

This commit additionally includes a minor version bump for the
wit-parser crate.