Include docs in imports and exports.
Include the parsed documentation comment in the AST's `Import` and `Export`
records so that, for example, documentation for the following function
can be included in the generated bindings.

```
default world component {
    /// Say hello!
    export hello-world: func() -> string
}
```

This will make the documentation generated by `cargo-component doc`
added in bytecodealliance/cargo-component#63 more useful, and will
allow me to fix a bug in a test which is supposed to be testing that the
documentation comments show up.