Implement structural equivalence for interfaces when merging worlds.
When merging two worlds together, coalesce interfaces that are
structurally equivalent to each other.

The motivation for this is that I want to be able to write library code
that uses wit bindings that can be used in any world that contains that
interface.

One such library is std, which will want to depend on wasi-filesystem
and other interfaces, but remain independent of the specific world that
the user will ultimately chose to link everything into.