Fix nesting order of rec/sub definitions
The text format is not completely specified for type
definitions, but there's current agreement that the
nesting of constructs goes:

(rec (type (sub (func/struct/array)))

'rec' and 'sub' may be omitted, but 'type' may not.

This commit fixes the text parser to follow this.
Previously it went rec -> sub -> type.