wast: Update GC proposal instructions (#972)
* Implement the GC proposal bottom types

The latest version of the GC proposal has a
'bottom' heap type for each type hierarchy
any, func, and extern. It's purpose is to be
a type for nulls in each hierarchy.

* Update GC instructions to latest proposal snapshot

This commit updates the GC proposal instructions to the
V8 snapshot version 6 [1].

New instructions:
  1. array.fill, init_data, init_elem
  2. refactored casting instructions
    - ref.cast/test/br_on_cast now take reftypes
    - ref.as_HEAP_TYPE are gone

[1] https://docs.google.com/document/d/1DklC3qVuOdLHSXB5UXghM_syCh-4cMinQ50ICiXnK3Q/edit#heading=h.rz955hp0nqsw

* Fix nesting order of rec/sub definitions

The text format is not completely specified for type
definitions, but there's current agreement that the
nesting of constructs goes:

(rec (type (sub (func/struct/array)))

'rec' and 'sub' may be omitted, but 'type' may not.

This commit fixes the text parser to follow this.
Previously it went rec -> sub -> type.