Add a bit extra fuel in the mutated store
If an original wasm traps then it is not guaranteed to perfectly reflect
the amount of fuel needed to reach the trap. For example the fuel
counter isn't flushed out except on function calls or other
synchronizing style events. At this time instructions which trap aren't
considered "barriers" in this sense which means the mutated store may
require a bit more fuel than what's reported to reach the same trap.