Implement validation of imported/exported type names
This commit implements a long-desired feature for wasmparser's
implementation of the component model, which is to validate that all
referred-to-types in imports and exports are all themselves named via
imports or exports. This design took a surprising amount of
back-and-forth to settle on something which felt reasonable enough to
land, but the general idea landed on in this commit is relatively
simple:

* The validator's component state now maintains a set of
  exported/imported types.
* When exporting/importing an item, all referred-to types must be
  present in these sets.
* When exporting/importing a type, it's added to these sets.
* When exporting/importing an instance, that recursively is reached into
  to additionally add to these sets.

A minor optimization is that when validating that all names are in the
appropriate set it only has to travel one layer deep in the type
hierarchy since if the one-layer-deep version is "in the set" then it's
already previously validated that everything else is in the set.

The validation predicate here is done with sets and type traversal so
it's probably no the fastest thing in the world, but after all the
back-and-forth about how to implement this I think this is a reasonable
place to start and continue to iterate on the future as necessary.

It's important to note that while this does not alter the binary format
this does change the validation algorithm meaning previously valid
components may no longer be valid. For example the old output of
`wit-component` is no longer considered valid for exported instances.
Instead the output has been updated here to ensure the generated
components conform to the new shape expected by wasmparser.

The other affected crates in this commit are:

* `wit-component` - construction of exported instances is updated so the
  imports in to the shim-component generated are no longer aliased from
  the outer context but redefined in the shim's context. This bloats
  components a bit but the type information additionally shouldn't be
  that intensive.

* `wasm-smith` - the test for generated components has been disabled.
  This would need an update to filter what can be used in exports, but
  for now `wasm-smith` has fallen far enough behind the current state of
  components it seemed more prudent to simply disable it and get back to
  it later.