Fix a panic in `wast` with components (#983)
This commit fixes an issue in the `wast` crate where an inline export
could be listed on a type definition in a type context in the component
model. This caused a panic when converting to binary due to the inline
export being unexpected.

This resolves the issue by disallowing inline exports in this context,
instead only allowing it in the top-level component context.