Initial implementation of kebab names for resources
This commit is an initial pass at implementing methods, constructors,
and static methods on resource types. These introduce names of the
form:

* `[method]foo.bar`
* `[constructor]foo`
* `[static]foo.baz`

which all have their own set of validation rules about when the names
can be mentioned and when they can be applied. The current iteration of
validation and design is working through some upstream questions at this
time but otherwise the implementation here was relatively
straightforward and largely changing some base types and then
propagating the change outwards.